Initial definition of application
This commit adds an extremely small amount of logic as a "base" for the
application. It is broken down into two components:

- A "interactive" component (goproctl)
- A "daemon" component (goprod)

The daemon takes the responsibility of watching for the camera to be
unplugged or plugged in or so, and setting up the required connections.
When streaming video, gopro requires there to be a "keepalive" as well
to keep the connection alive — the daemon would be responsible for this.

The interactive component would be to start, stop and otherwise interact
with the gopro through a CLI interface.

== Design Notes
=== This is planned to be an iteration of the "oilr project" golang
tooling, designed for littleman.co