Initial implementation of "hotplug" feature
This commit drives the initial implementation of the "hotplug" feature —
essentially a thing that watches for when a GoPro (with the firmware
that supports being treated as an ethernet device) is connected.

Futre work will build on this to either automatically or by command,
start the cameras "streaming" operation.

== Design Notes
=== API Design

This work is a hobby, and thus not of a particularly excellent quality.
However, the general desire for the author was to have a good split
between the packages such that no matter the craziness that was
happening within the page, the context outside the package would see
only a fairly clean design.

=== Arbitrary GoPro features

Strangely, it appears that there are certain constants (for example, IP)
that the GoPro presents itself. They have thus been hard-coded into the
program.

It is not yet clear whether these are hard coded or derived from some
property of the camera embedded when the camera is manufactured or just
consistent across all cameras.

Naively the author would have expected some sort of discovery protocol
(that would also work over Wifi), but they havent looked into this
deeply yet.

== Limitations
=== No Telemetry

Initially the author planned to do some work to implement OpenTelemetry
in the various libraries.

The author just wants their shiny thing, and thus will skip this for
now.

=== Doesn't scan on boot

At the time of writing this doesn't scan USB devices that were already
booted, and thus misses the camera if it was already connected.

Future work should implement this scan when Listen() is called, stubbing
out the connection events.