Switch the webcam on automatically on connect
This project has, broadly speaking, taken an absolute dive in quality.
I just want to get "something" useful — the same as I would in a bash
script — up as quickly as I can now, and then tidy it up "at some future
point".

That future point may or may not come. I'm not sure whether the gopro
will be valuable yet, as a webcam. Probably?